requirements.txt
openai>=2.2.0
langgraph>=1.0.2
pydantic>=2.6.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
